x64: fix wx3 runtime asserts (CharDetailsFrame stall, format specifier)
wxWidgets 3.x keeps asserts enabled in release builds and shows a modal
dialog by default, which blocked the headless run while constructing
CharDetailsFrame. Add wxApp::OnAssertFailure to log and continue instead
of popping a dialog (also the right behaviour for end users), then fix the
two underlying asserts: CharDetailsFrame combined wxSizerFlags Expand()
with Align() (illegal in wx3 - drop the alignment), and lightcontrol fed a
size_t (64-bit on x64) to a %i format specifier (cast to int).

Verified: the x64 build launches, builds the wx3 UI assert-free, and loads
the full game data correctly (ChrRaces 0-changed vs curated schema, 110048
items, 130572 files) - matching the 32-bit build.

+++ Source/wowmodelviewer/lightcontrol.cpp
@@ -70,7 +70,7 @@ LightControl::LightControl(wxWindow* parent, wxWindowID id)
   wxArrayString choices;
   for (size_t i=1; i<=MAX_LIGHTS; i++) {
     wxString s = wxT("Light ");
-    s += wxString::Format(wxT("%i"), i);
+    s += wxString::Format(wxT("%i"), (int)i);
 
     choices.Add(s);
   }
